Update fake_file_system.py
Changing how the starting directory is added to the fake file system. Adding the ability to save an instance of a fake file system then load it back into a running session later. Removing direct logging due to errors going to be more common for issue acknowledgment in future updates